+++ tests/docapi_file_service_test.py
@@ -1,9 +1,8 @@
 import unittest
-import uuid
 
 from vvrest.services.doc_api_file_service import DocApiFileService
 
-from .utilities import get_vault_object, get_jwt, get_parameters_json
+from .utilities import get_vault_object, get_parameters_json
 
 
 class DocApiFileServiceTest(unittest.TestCase):
@@ -17,7 +16,7 @@ class DocApiFileServiceTest(unittest.TestCase):
         test_parameters = get_parameters_json()
         cls.document_id = test_parameters['document_id']
         cls.document_revision_id = test_parameters['document_revision_id']
-        cls.jwt = get_jwt(cls.vault)
+        cls.vault_jwt = get_vault_object(user_web_token=None, jwt=None, auto_jwt=True)
 
     def test_docapi_file_not_enabled(self):
         """
@@ -32,8 +31,7 @@ class DocApiFileServiceTest(unittest.TestCase):
         """
         validates 404 is returned if not a valid dhid
         """
-        vault = get_vault_object(user_web_token=None, jwt=self.jwt)
-        file_service = DocApiFileService(vault)
+        file_service = DocApiFileService(self.vault_jwt)
         resp = file_service.get_file_stream(self.document_id)
         self.assertEqual(resp.status_code, 404)
 
@@ -41,8 +39,7 @@ class DocApiFileServiceTest(unittest.TestCase):
         """
         validates successful doc api file stream call
         """
-        vault = get_vault_object(user_web_token=None, jwt=self.jwt)
-        file_service = DocApiFileService(vault)
+        file_service = DocApiFileService(self.vault_jwt)
         resp = file_service.get_file_stream(self.document_revision_id)
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(resp.text, 'test file')