+++ tests/utilities.py
@@ -3,22 +3,22 @@ from uuid import uuid4
 from random import choice
 
 from vvrest.vault import Vault
-from vvrest.services.user_service import UserService
 
 from .settings import credentials_file, parameters_file
 
 
-def get_vault_object(user_web_token=None, jwt=None):
+def get_vault_object(user_web_token=None, jwt=None, auto_jwt=False):
     """
     :param user_web_token: string UUID(version=4), used for user impersonation
     :param jwt: str, JSON Web Token
+    :param auto_jwt: bool
     :return: Vault
     """
     with open(credentials_file) as credentials_json:
         credentials = json.load(credentials_json)
 
     vault = Vault(credentials['url'], credentials['customer_alias'], credentials['database_alias'],
-                  credentials['client_id'], credentials['client_secret'], user_web_token, jwt)
+                  credentials['client_id'], credentials['client_secret'], user_web_token, jwt, auto_jwt)
 
     return vault
 
@@ -61,13 +61,3 @@ def get_random_string(length):
     random_string = ''.join(choice(char_choices) for i in range(length))
 
     return random_string
-
-
-def get_jwt(vault):
-    """
-    fetches jwt for a user
-    """
-    user_service = UserService(vault)
-    resp = user_service.get_user_jwt()
-
-    return resp['data']['token']